Create helpers folder
refactor: Create helpers folder for component helper functions

- Create helpers folder and filterVariant in it

- Migrate all the helper functions for FilteredVariant to filterVariant
